Move export symbol methods to ModuledNode from SourceFile.
